requirements.txt
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.0.2
Flask-RESTful==0.3.8
Jinja2==2.11.3
MarkupSafe==1.1.1
PTable==0.9.2
PyJWT==2.0.1
Schemer==0.2.11
Schemer3==0.2.11.post1
Werkzeug==1.0.1
aiohttp==3.6.3
aniso8601==9.0.0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.2.0
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
codecov==2.1.10
colorama==0.4.4
coverage==5.3
fpl==0.6.25
idna==2.10
inflection==0.2.0
iniconfig==1.1.1
itsdangerous==1.1.0
mongoengine==0.22.1
multidict==4.7.6
packaging==20.4
pip==20.2.1
pluggy==0.13.1
py==1.9.0
pycparser==2.20
pydash==4.9.2
pymongo==3.11.3
pyparsing==2.4.7
pytest==6.1.1
pytest-aiohttp==0.3.0
pytest-cov==2.10.1
pytest-mock==3.3.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.24.0
setuptools==49.2.1
six==1.15.0
toml==0.10.1
urllib3==1.25.10